Server/suber/requirements.txt
Django==3.1.5
djangorestframework==3.11.2
requests==2.25.1